+++ packages/core/tests/factories/factory.spec.ts
@@ -14,14 +14,21 @@ const TestFactory = new Factory<IService>("service");
  */
 class TestService implements IService {
 	/**
+	 * The name of the service.
+	 */
+	public readonly name: string;
+
+	/**
 	 * Validation property.
 	 */
 	public foo: number;
 
 	/**
 	 * Create a new instance of TestService.
+	 * @param name The name of the service.
 	 */
-	constructor() {
+	constructor(name: string) {
+		this.name = name;
 		this.foo = 1;
 	}
 
@@ -72,7 +79,7 @@ describe("Factory", () => {
 	});
 
 	test("register and get can succeed", () => {
-		TestFactory.register("test", () => new TestService());
+		TestFactory.register("test", () => new TestService("test"));
 		const t = TestFactory.get<TestService>("test");
 		expect(t).toBeDefined();
 		if (t) {
@@ -81,7 +88,7 @@ describe("Factory", () => {
 	});
 
 	test("register and get can succeed and return same instance", () => {
-		const testService = new TestService();
+		const testService = new TestService("test2");
 		TestFactory.register("test2", () => testService);
 		const t2 = TestFactory.get<TestService>("test2");
 		if (t2) {
@@ -96,7 +103,7 @@ describe("Factory", () => {
 	});
 
 	test("unregister can succeed", () => {
-		TestFactory.register("test3", () => new TestService());
+		TestFactory.register("test3", () => new TestService("test3"));
 		const t3 = TestFactory.get<TestService>("test3");
 		if (t3) {
 			expect(t3.foo).toEqual(1);
@@ -129,18 +136,68 @@ describe("Factory", () => {
 	});
 
 	test("can reset the factory", () => {
-		TestFactory.register("test3", () => new TestService());
+		TestFactory.register("test3", () => new TestService("test3"));
 		expect(TestFactory.get<TestService>("test3").foo).toEqual(1);
 		TestFactory.get<TestService>("test3").foo = 2;
 		TestFactory.reset();
 		expect(TestFactory.get<TestService>("test3").foo).toEqual(1);
 	});
 
+	test("can auto register services", () => {
+		const testFactory = new Factory<TestService>("service1", true);
+		testFactory.register("test1", () => new TestService("test1"));
+		const instanceList = testFactory.instancesList();
+		expect(instanceList.length).toEqual(1);
+		expect(instanceList[0].name).toEqual("test1");
+	});
+
+	test("can register multiple services and get them as an array ordered by addition order", () => {
+		const testFactory = new Factory<TestService>("service1");
+		testFactory.register("test3", () => new TestService("test3"));
+		testFactory.register("test2", () => new TestService("test2"));
+		testFactory.register("test1", () => new TestService("test1"));
+		testFactory.get("test1");
+		testFactory.get("test2");
+		testFactory.get("test3");
+		const instanceList = testFactory.instancesList();
+		expect(instanceList.length).toEqual(3);
+		expect(instanceList[0].name).toEqual("test3");
+		expect(instanceList[1].name).toEqual("test2");
+		expect(instanceList[2].name).toEqual("test1");
+	});
+
+	test("can register multiple services and get them as a map", () => {
+		const testFactory = new Factory<TestService>("service1");
+		testFactory.register("test3", () => new TestService("test3"));
+		testFactory.register("test2", () => new TestService("test2"));
+		testFactory.register("test1", () => new TestService("test1"));
+		testFactory.get("test1");
+		testFactory.get("test2");
+		testFactory.get("test3");
+		const instanceMap = testFactory.instancesMap();
+		expect(Object.keys(instanceMap).length).toEqual(3);
+		expect(instanceMap.test3.name).toEqual("test3");
+		expect(instanceMap.test2.name).toEqual("test2");
+		expect(instanceMap.test1.name).toEqual("test1");
+	});
+
+	test("can register multiple services and get their names as an array ordered by addition order", () => {
+		const testFactory = new Factory<TestService>("service1");
+		testFactory.register("test3", () => new TestService("test3"));
+		testFactory.register("test2", () => new TestService("test2"));
+		testFactory.register("test1", () => new TestService("test1"));
+		const names = testFactory.names();
+		expect(names.length).toEqual(3);
+		expect(names[0]).toEqual("test3");
+		expect(names[1]).toEqual("test2");
+		expect(names[2]).toEqual("test1");
+	});
+
 	test("can have two factories which don't collide", () => {
 		const testFactory1 = new Factory<IService>("service1");
 		const testFactory2 = new Factory<IService>("service2");
-		testFactory1.register("test1", () => new TestService());
-		testFactory2.register("test2", () => new TestService());
+		testFactory1.register("test1", () => new TestService("test1"));
+		testFactory2.register("test2", () => new TestService("test2"));
 
 		const names1 = testFactory1.names();
 		expect(names1.length).toEqual(1);
@@ -149,5 +206,15 @@ describe("Factory", () => {
 		const names2 = testFactory2.names();
 		expect(names2.length).toEqual(1);
 		expect(names2[0]).toEqual("test2");
+	});
+
+	test("can use a different name matching strategy", () => {
+		const testFactory1 = new Factory<IService>("service1", false, (names, name) =>
+			names.find(n => `@${n}` === name)
+		);
+		testFactory1.register("test1", () => new TestService("test1"));
+
+		const instance = testFactory1.get("@test1");
+		expect(instance).toBeDefined();
 	});
 });