+++ packages/core/docs/changelog.md
@@ -6,6 +6,7 @@
 - Added: Base64Url class
 - Added: Converter.bytesToBase64Url class
 - Added: Converter.base64UrlToBytes class
+- Added: JsonHelper class
 
 ## v0.0.2